Add motor speed control signal derivative limiter.
Change from FIR to derivative limiter in step motor signal filtering.
Obtain experimentally maximum motor speed, first and second derivative values.

+++ STM32_firmware/Components/Encoder/Inc/encoder.h
@@ -34,7 +34,7 @@ typedef struct {
   int16_t Counter;
   int16_t CounterMax;
   int16_t CounterMin;
-  uint8_t Step;
+  uint16_t Step;
 } ENC_HandleTypeDef;
 
 /* Public function prototypes ------------------------------------------------*/
